обновил зависимости, настройка модальных окон через bootstrap

+++ src/app/services/storage/storage.service.ts
@@ -1,28 +1,50 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 
 @Injectable({
   providedIn: 'root'
 })
 export class StorageService {
 
-  constructor() { }
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) { }
 
   // localStorage methods
   setLocalItem(key: string, value: any): void {
-    localStorage.setItem(key, JSON.stringify(value));
+    if (isPlatformBrowser(this.platformId)){
+      localStorage.setItem(key, JSON.stringify(value));
+    }
+    else {
+      throw new Error('Ошибка доступа к localStorage');
+    }
   }
 
   getLocalItem(key: string): any {
-    const item = localStorage.getItem(key);
-    return item ? JSON.parse(item) : null;
+    if (isPlatformBrowser(this.platformId)){
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : null;
+    }
+    else {
+      throw new Error('Ошибка доступа к localStorage');
+      return null;
+    }
   }
 
   removeLocalItem(key: string): void {
-    localStorage.removeItem(key);
+    if (isPlatformBrowser(this.platformId)){
+      localStorage.removeItem(key);
+    }
+    else {
+      throw new Error('Ошибка доступа к localStorage');
+    }
   }
 
   clearLocalStorage(): void {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)){
+      localStorage.clear();
+    }
+    else {
+      throw new Error('Ошибка доступа к localStorage');
+    }
   }
 
   // sessionStorage methods
